band3 wave #6: port 14 meta_band TUs Wii->360 (compile-clean, +0 matches)
14 RB3 game-code TUs ported from rb3-Wii and wired NonMatching (objects.json
band3 meta_band: 29->43). All compile clean under MSVC X360; build green.
matched_functions holds at 394 (compile-only wiring — these are unpinned, so
no target .obj to diff against yet; matching is a later pass).

Ported: Accomplishment, AccomplishmentManager, AccomplishmentOneShot,
AccomplishmentProgress, Accomplishment{DiscSong,LessonDiscSong,LessonSongList,
Player,Song,SongFilter,SongList,Tour}Conditional, SongSortByDiff,
WiiBufStreamMgr.

Dropped (on disk, NOT wired): AccomplishmentPanel.cpp — 24 errs (PanelDir only
a stub, UIPanel::GetState const mismatch, switch-init). Committed as scaffold
for a future pass; inert (absent from objects.json, never compiled).

New MWCC->MSVC patterns (appended to meta_band-port-breaking-changes.md):
- _MemAlloc(int,int)/_MemFree fwd-decls added to MemMgr.h (STLPort specializations)
- kSuccess enum clash (PurchaseState vs JoinResponseError) -> forward-decl StorePurchaser
- operator new(unsigned long)->size_t (VirtualRootObject.h)
- TheContentMgr->m() ptr->ref; Achievements::Submit pad-num arg; String needs .c_str()
- strip #pragma dont_inline/pool_data (MWCC-only)

Shared-header touches (both additive/minimal, verified non-conflicting with the
concurrent codegen agent which made zero edits): MemMgr.h (+2 decls),
VirtualRootObject.h (size_t). New engine headers ported from rb3-Wii:
MeterDisplay.h, StoreArtLoaderPanel.h, UIGridProvider.h.

Build verified green via ./tools/ninja-locked (exit 0, all 14 .obj compiled).

+++ src/network/Platform/VirtualRootObject.h
@@ -3,7 +3,7 @@
 namespace Quazal {
     class VirtualRootObject {
     public:
-        static void *operator new(unsigned long, char *, unsigned int);
+        static void *operator new(size_t, char *, unsigned int); // size_t not unsigned long (MSVC requires)
         static void operator delete(void *);
     };
 }
